Read cache metadata file size correctly
Currently we will not be able to read cache metadata for files that are
bigger than 4GB because we use stoi function. Use stoll instead that
will read long long values.

+++ src/pgduckdb_options.cpp
@@ -264,7 +264,7 @@ DuckdbGetCachedFilesInfos() {
 				     p.path().c_str());
 				break;
 			}
-			cache_info.push_back(CacheFileInfo {metadata_tokens[0], metadata_tokens[1], std::stoi(metadata_tokens[2]),
+			cache_info.push_back(CacheFileInfo {metadata_tokens[0], metadata_tokens[1], std::stoll(metadata_tokens[2]),
 			                                    std::stoi(metadata_tokens[3])});
 		}
 	}
